+++ packages/waku/src/lib/plugins/vite-plugin-nonjs-resolve.ts
